Add missing tests to unit tests
