--- addons/Dexie.Syncable/test/test-syncable.html
@@ -1,24 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>Dexie.Syncable Unit tests</title>
-    <link rel="stylesheet" href="../../../node_modules/qunitjs/qunit/qunit.css">
-</head>
-<body>
-    <div id="qunit"></div>
-    <div id="qunit-fixture"></div>
-    <script src="../../../node_modules/qunitjs/qunit/qunit.js"></script>
-    <script src="../../../dist/dexie.js"></script>
-    <script src="../../Dexie.Observable/dist/dexie-observable.js"></script>
-    <script src="../dist/dexie-syncable.js"></script>
-    <script src="../../../samples/remote-sync/websocket/WebSocketSyncProtocol.js"></script>
-    <script src="../../../samples/remote-sync/websocket/websocketserver-shim.js"></script>
-    <script src="../../../samples/remote-sync/websocket/WebSocketSyncServer.js"></script> <!-- With shim applied, we can run the server in the browser -->
-    <script src="tests-WebSocketSyncServer.js"></script>
-    <script src="tests-syncable.js"></script>
-    <script src="tests-syncable-partials.js"></script>
-    <script src="tests-syncprovider.js"></script>
-    <script src="tests-changing-options.js"></script>
-</body>
-</html>
