+++ addons/Dexie.Syncable/test/unit/karma.conf.js
@@ -15,6 +15,10 @@ module.exports = function (config) {
     files: karmaCommon.files.concat([
       'dist/dexie.js',
       'addons/Dexie.Observable/dist/dexie-observable.js',
+
+      'samples/remote-sync/websocket/websocketserver-shim.js',
+      'samples/remote-sync/websocket/WebSocketSyncServer.js',// With shim applied, we can run the server in the browser
+
       'addons/Dexie.Syncable/test/unit/bundle.js',
       { pattern: 'addons/Dexie.Observable/dist/*.map', watched: false, included: false },
       { pattern: 'addons/Dexie.Syncable/dist/*.map', watched: false, included: false },
