+++ addons/Dexie.Syncable/test/unit/run-unit-tests.html
@@ -13,6 +13,8 @@
     <script src="../../../../dist/dexie.js"></script>
     <script src="../../../Dexie.Observable/dist/dexie-observable.js"></script>
     <script src="../../dist/dexie-syncable.js"></script>
+    <script src="../../../../samples/remote-sync/websocket/websocketserver-shim.js"></script>
+    <script src="../../../../samples/remote-sync/websocket/WebSocketSyncServer.js"></script> <!-- With shim applied, we can run the server in the browser -->
     <script src="bundle.js"></script>
 </body>
 </html>
