+++ addons/Dexie.Syncable/test/unit/tests-PersistedContext.js
@@ -1,12 +1,12 @@
 import Dexie from 'dexie';
-import 'dexie-observable';
+import observable from 'dexie-observable';
 // Add this so we have the SyncNode.prototype.save method
-import '../../src/Dexie.Syncable';
+import syncable from '../../src/Dexie.Syncable';
 import {module, asyncTest, test, start, stop, propEqual, deepEqual, ok} from 'QUnit';
 import {resetDatabase} from '../../../../test/dexie-unittest-utils';
 import initPersistedContext from '../../src/PersistedContext';
 
-const db = new Dexie('TestDBTable');
+const db = new Dexie('TestDBTable', {addons: [observable, syncable]});
 db.version(1).stores({foo: '++id'});
 
 module('PersistedContext', {
