+++ addons/Dexie.Syncable/test/unit/unit-tests-all.js
@@ -3,6 +3,7 @@ import './get-local-changes-for-node/tests-get-changes-since-revision.js';
 import './get-local-changes-for-node/tests-get-local-changes-for-node.js';
 import './tests-apply-changes.js';
 import './tests-bulk-update.js';
+import './tests-changing-options.js';
 import './tests-combine-create-and-update.js';
 import './tests-combine-update-and-update.js';
 import './tests-finally-commit-all-changes.js';
@@ -11,3 +12,7 @@ import './tests-merge-change.js';
 import './tests-PersistedContext.js';
 import './tests-register-sync-protocol.js';
 import './tests-save-to-uncommitted-changes.js';
+import './tests-syncable.js';
+import './tests-syncable-partials.js';
+import './tests-syncprovider.js';
+import './tests-WebSocketSyncServer.js';
